Refactor createEmptyData
