+++ packages/canner-helpers/src/createEmptyData.js
@@ -14,66 +14,9 @@
  */
 
 import invariant from 'invariant';
+import {mapSchema, genDefaultValue} from './utils';
+import type {Schema, SchemaMap} from './types';
 
-type ObjectSchema = {
-  items: SchemaMap;
-  type: 'object';
-  ui: string;
-}
-
-type ArraySchema = {
-  items: Schema;
-  type: 'array';
-  ui: string;
-}
-
-type StringSchema = {
-  type: 'string';
-  ui: string;
-}
-
-type NumberSchema = {
-  type: 'number';
-  ui: string;
-}
-
-type BooleanSchema = {
-  type: 'boolean';
-  ui: string;
-}
-
-type RelationSchema = {
-  type: 'relation';
-  ui: string;
-  relation: {
-    type: string;
-    to: string;
-  }
-};
-
-type GeoPointSchema = {
-  type: 'geoPoint',
-
-}
-
-type DateTimeSchema = {
-  type: 'dateTime'
-}
-
-type FileSchema = {
-  type: 'file',
-}
-
-type ImageScheme = {
-  type: 'image'
-}
-
-type Schema = ArraySchema | ObjectSchema | StringSchema | BooleanSchema | NumberSchema | RelationSchema
-  | GeoPointSchema | DateTimeSchema | FileSchema | ImageScheme;
-
-type SchemaMap = {
-  [string]: Schema;
-};
 
 function loop(schema: Schema) {
   let result: any;
@@ -83,7 +26,7 @@ function loop(schema: Schema) {
   }
   switch (schema.type) {
     case 'object':
-      result = mapSchema(schema.items);
+      result = mapSchema(schema.items, loop);
       result.__typename = result.__typename || null;
       break;
     case 'array':
@@ -151,21 +94,6 @@ function loop(schema: Schema) {
   return result;
 }
 
-function mapSchema(schemaMap: SchemaMap) {
-  return Object.keys(schemaMap || {})
-    .reduce((result: {[string]: mixed}, key: string) => {
-      result[key] = loop(schemaMap[key]);
-      return result;
-    }, {});
-}
-
-export function genDefaultValue(defaultValue: any) {
-  if (typeof defaultValue === 'function') {
-    return defaultValue();
-  }
-  return defaultValue;
-}
-
 export default function(schema: Schema | SchemaMap) {
   if (!schema) {
     return null;
@@ -173,5 +101,5 @@ export default function(schema: Schema | SchemaMap) {
   if ('type' in schema) {
     return loop(((schema: any): Schema));
   }
-  return mapSchema(((schema: any): SchemaMap));
+  return mapSchema(((schema: any): SchemaMap), loop);
 }
